pyproject.toml
[project]
name = "koreo"
version = "0.1.0"
description = "Type-safe and testable KRM Templates and Workflows."
authors = [
    {name = "Robert Kluin", email = "robert.kluin@realkinetic.com"},
]

dependencies = [
    "cel-python==0.1.5",
    "kopf==1.37.4",
    "jsonpath-ng==1.7.0",
    "PyYAML==6.0.2",
    "kr8s==0.18.1",
    "uvloop>=0.20.0",
]

requires-python = ">=3.13"
readme = "README.md"
license = {text = "Apache-2.0"}

[project.urls]
Homepage = "https://realkinetic.com/konfig"

[project.optional-dependencies]
dev = [
    "black>=24.8.0",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [
  "-v",
  "--import-mode=importlib",
  "--cov=src",
  "--cov-branch",
  "--cov-report=term-missing",
]
